+++ clear-folio-maven-cache.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-
-rm -rf ~/.m2/repository/org/folio
-